Fixed failed to move
Added error handling: I added a try and except block around the shutil.move() function. This ensures that if an error occurs while moving a file, it will be caught and the error message will be printed, instead of causing the script to crash. This change doesn't necessarily cause the files to be moved successfully, but it helps to identify issues that might arise during the move process.

Implemented a counter to track successful file moves: Instead of comparing the contents of the source and destination folders directly, I added a counter (successful_moves) that increments for each successful file move. This counter is compared to the total number of files in the source folder (total_files). If the number of successful moves is equal to the total number of files, it means all files were moved successfully, and the source folder can be safely deleted.